Move events to their own dir, fix more tests
