+++ src/Events/ExecEvent.php
@@ -6,7 +6,7 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace Yiisoft\Yii\Queue;
+namespace Yiisoft\Yii\Queue\Events;
 
 /**
  * Exec Event.
