+++ tests/JobEventTest.php
@@ -9,11 +9,11 @@
 namespace Yiisoft\Yii\Queue\Tests;
 
 use Yiisoft\Yii\Queue\Closure\Behavior as ClosureBehavior;
-use Yiisoft\Yii\Queue\ExecEvent;
+use Yiisoft\Yii\Queue\Events\ExecEvent;
 use Yiisoft\Yii\Queue\InvalidJobException;
-use Yiisoft\Yii\Queue\JobEvent;
-use Yiisoft\Yii\Queue\Queue;
+use Yiisoft\Yii\Queue\Events\JobEvent;
 use Yiisoft\Yii\Queue\Drivers\Sync\Queue as SyncQueue;
+use Yiisoft\Yii\Queue\Serializers\JsonSerializer;
 
 /**
  * Job Event Test.
@@ -28,25 +28,27 @@ class JobEventTest extends TestCase
         $eventHandler = function (JobEvent $event) use (&$eventCounter) {
             $eventCounter[$event->id][$event->name] = true;
         };
-        $queue = new SyncQueue(['strictJobType' => false]);
-        $queue->on(Queue::EVENT_BEFORE_EXEC, $eventHandler);
-        $queue->on(Queue::EVENT_AFTER_ERROR, $eventHandler);
-        $queue->on(Queue::EVENT_AFTER_ERROR, function (ExecEvent $event) {
+        $queue = new SyncQueue(new JsonSerializer());
+        $queue->strictJobType = false;
+        $queue->on(ExecEvent::BEFORE, $eventHandler);
+        $queue->on(ExecEvent::AFTER, $eventHandler);
+        $queue->on(ExecEvent::AFTER, function (ExecEvent $event) {
             $this->assertTrue($event->error instanceof InvalidJobException);
             $this->assertFalse($event->retry);
         });
         $jobId = $queue->push('message that cannot be unserialized');
         $queue->run();
         $this->assertArrayHasKey($jobId, $eventCounter);
-        $this->assertArrayHasKey(Queue::EVENT_BEFORE_EXEC, $eventCounter[$jobId]);
-        $this->assertArrayHasKey(Queue::EVENT_AFTER_ERROR, $eventCounter[$jobId]);
+        $this->assertArrayHasKey(ExecEvent::BEFORE, $eventCounter[$jobId]);
+        $this->assertArrayHasKey(ExecEvent::AFTER, $eventCounter[$jobId]);
     }
 
     public function testExecResult()
     {
-        $queue = new SyncQueue(['as closure' => ClosureBehavior::class]);
+        $queue = new SyncQueue(new JsonSerializer());
+        $queue->attachBehavior('closure', ClosureBehavior::class);
         $isTriggered = false;
-        $queue->on(Queue::EVENT_AFTER_EXEC, function (ExecEvent $event) use (&$isTriggered) {
+        $queue->on(ExecEvent::AFTER, function (ExecEvent $event) use (&$isTriggered) {
             $isTriggered = true;
             $this->assertSame(12345, $event->result);
         });
