+++ tests/app/config/console.php
@@ -26,6 +26,6 @@ return [
                 'Yiisoft\Yii\Queue\Drivers\Db\migrations',
             ],
         ],
-        'benchmark' => \Yiisoft\Yii\Queue\Tests\App\Benchmark\Controller::class,
+        'benchmark' => \Yiisoft\Yii\Queue\Tests\app\benchmark\Controller::class,
     ],
 ];
