+++ tests/drivers/db/TestCase.php
@@ -9,8 +9,8 @@
 namespace Yiisoft\Yii\Queue\Tests\drivers\db;
 
 use yii\db\Query;
-use Yiisoft\Yii\Queue\Tests\App\PriorityJob;
-use Yiisoft\Yii\Queue\Tests\App\RetryJob;
+use Yiisoft\Yii\Queue\Tests\app\PriorityJob;
+use Yiisoft\Yii\Queue\Tests\app\RetryJob;
 use Yiisoft\Yii\Queue\Tests\drivers\CliTestCase;
 
 /**
